+++ crates/cubek-matmul/src/components/mod.rs
@@ -1,9 +1,7 @@
 pub mod batch;
 pub mod global;
 pub mod stage;
-pub mod tile_matmul;
-
-mod resource;
+pub mod tile;
 
 // Internal-only — external crates import this directly from cubek-std.
-pub(crate) use resource::*;
+pub(crate) use cubek_std::CubeDimResource;
