--- crates/cubek-matmul/src/components/resource.rs
@@ -1,1 +1,0 @@
-pub use cubek_std::CubeDimResource;
