+++ crates/cubek-matmul/src/components/stage/matmul/plane_partitioned/matmul.rs
@@ -1,8 +1,8 @@
 use crate::{
     components::global::PlaneFlowPartition, components::global::PlaneFlowPartitionRule,
     components::stage::matmul::partitioned_matmul::PartitionedStageMatmul,
-    components::stage::matmul::partitioned_matmul::StagePartitioner,
-    components::tile_matmul::Plane, definition::MatmulTypes,
+    components::stage::matmul::partitioned_matmul::StagePartitioner, components::tile::Plane,
+    definition::MatmulTypes,
 };
 use cubecl::{prelude::*, std::tensor::layout::Coords2d};
