+++ crates/cubek-matmul/src/routines/selector/plane.rs
@@ -19,7 +19,7 @@ use crate::routines::selector::is_tiny;
 use crate::{
     components::global::{InputLoadFlow, LoadFlows},
     components::stage::PartitionBuffering,
-    components::tile_matmul::TileMatmulKind,
+    components::tile::TileMatmulKind,
 };
 
 pub const NUM_SM_APPROX: u32 = 50;
